+++ packages/server/api/src/app/ee/database/migrations/postgres/1686133672743-AddPinnedAndBlogUrlToTemplates.ts
@@ -1,10 +1,15 @@
 import { MigrationInterface, QueryRunner } from 'typeorm'
+import { isNotOneOfTheseEditions } from '../../../../database/database-common'
+import { ApEdition } from '@activepieces/shared'
 
 export class AddPinnedAndBlogUrlToTemplates1686133672743
 implements MigrationInterface {
     name = 'AddPinnedAndBlogUrlToTemplates1686133672743'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.ENTERPRISE, ApEdition.CLOUD])) {
+            return
+        }
         await queryRunner.query(
             'ALTER TABLE "flow_template" ADD "pinned" boolean NOT NULL',
         )
@@ -14,6 +19,9 @@ implements MigrationInterface {
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.ENTERPRISE, ApEdition.CLOUD])) {
+            return
+        }
         await queryRunner.query(
             'ALTER TABLE "flow_template" DROP COLUMN "blogUrl"',
         )
