+++ packages/server/api/src/app/ee/database/migrations/postgres/1689806173642-RemoveCalculatedMetrics.ts
@@ -1,10 +1,15 @@
 import { MigrationInterface, QueryRunner } from 'typeorm'
+import { isNotOneOfTheseEditions } from '../../../../database/database-common'
+import { ApEdition } from '@activepieces/shared'
 
 export class RemoveCalculatedMetrics1689806173642
 implements MigrationInterface {
     name = 'RemoveCalculatedMetrics1689806173642'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.CLOUD, ApEdition.ENTERPRISE])) {
+            return
+        }
         await queryRunner.query(
             'ALTER TABLE "project_usage" DROP COLUMN "activeFlows"',
         )
@@ -17,6 +22,9 @@ implements MigrationInterface {
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.CLOUD, ApEdition.ENTERPRISE])) {
+            return
+        }
         await queryRunner.query(
             'ALTER TABLE "project_usage" ADD "teamMembers" integer NOT NULL',
         )
