+++ packages/server/api/src/app/ee/database/migrations/postgres/1694604120205-AddFeaturedDescriptionAndFlagToTemplates.ts
@@ -1,10 +1,15 @@
 import { MigrationInterface, QueryRunner } from 'typeorm'
+import { isNotOneOfTheseEditions } from '../../../../database/database-common'
+import { ApEdition } from '@activepieces/shared'
 
 export class AddFeaturedDescriptionAndFlagToTemplates1694604120205
 implements MigrationInterface {
     name = 'AddFeaturedDescriptionAndFlagToTemplates1694604120205'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.ENTERPRISE, ApEdition.CLOUD])) {
+            return
+        }
         await queryRunner.query(
             'ALTER TABLE "flow_template" ADD "isFeatured" boolean',
         )
@@ -14,6 +19,9 @@ implements MigrationInterface {
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
+        if (isNotOneOfTheseEditions([ApEdition.ENTERPRISE, ApEdition.CLOUD])) {
+            return
+        }
         await queryRunner.query(
             'ALTER TABLE "flow_template" DROP COLUMN "featuredDescription"',
         )
